add check only when submit happens
